DE/requirements.txt
selenium==4.29.0
requests==2.31.0
azure-storage-blob==12.19.0
azure-identity==1.21.0
python-dotenv==1.1.0
pandas==2.2.3